fix: weird api error
