+++ tsconfig.json
@@ -34,12 +34,9 @@
       "@payload-config": [
         "./src/payload.config.ts"
       ],
-      "react": [
-        "./node_modules/@types/react"
-      ],
       "@/*": [
         "./src/*"
-      ],
+      ]
     }
   },
   "include": [
